refactor: PieceBuilder 추가
