+++ src/main/java/chess/domain/board/Builder.java
@@ -1,0 +1,9 @@
+package chess.domain.board;
+
+import chess.domain.piece.Piece;
+
+import java.util.Map;
+
+public interface Builder {
+    Map<Position, Piece> build();
+}
